# public/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FileFlow Project API NEW",
    "version": "1.0.0",
    "description": "API documentation for the FileFlow Project. All endpoints are synchronized with backend routes as of September 27, 2025."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/files/user-files": {
      "get": {
        "summary": "List User Files",
        "description": "Returns a list of files for the authenticated user.",
        "responses": {
          "200": {
            "description": "List of files returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/user-files/bulk-download": {
      "post": {
        "summary": "Bulk Download Files",
        "description": "Downloads multiple files as a ZIP archive.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "fileIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ZIP archive returned."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Files not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/user-files/reorder": {
      "post": {
        "summary": "Reorder User Files",
        "description": "Saves the new order of user files.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "order"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order saved successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/user-files/{fileId}/annotation": {
      "patch": {
        "summary": "Update File Annotation",
        "description": "Updates the annotation for a specific file.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "annotation": {
                    "type": "string"
                  }
                },
                "required": [
                  "annotation"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Annotation updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/user-files/{fileId}/favorite": {
      "post": {
        "summary": "Toggle Favorite",
        "description": "Favorites or unfavorites a file.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite status updated."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/user-files/{fileId}/metadata": {
      "patch": {
        "summary": "Update File Metadata",
        "description": "Updates metadata (caption, tags) for a file.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metadata updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/user-files/{fileId}/share": {
        "post": {
          "summary": "Share File",
          "description": "Shares a file with another user or generates a shareable link.",
          "parameters": [
            {
              "name": "fileId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "shareWithUserId": {
                      "type": "string",
                      "description": "User ID to share the file with (optional if generating a link)"
                    },
                    "generateLink": {
                      "type": "boolean",
                      "description": "Whether to generate a shareable link"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "description": "Expiration time for the share link in seconds (optional)"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "responses": {
          "200": {
            "description": "File shared successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/transform/{fileId}": {
        "post": {
          "summary": "Transform File",
          "description": "Transforms a file (e.g., image processing).",
          "parameters": [
            {
              "name": "fileId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "operation": {
                      "type": "string",
                      "description": "Type of transformation (e.g., resize, crop, convert)"
                    },
                    "options": {
                      "type": "object",
                      "description": "Transformation options (e.g., width, height, format)",
                      "additionalProperties": true
                    }
                  },
                  "required": ["operation"]
                }
              }
            }
          },
          "responses": {
          "200": {
            "description": "File transformed successfully."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/user-files/{fileId}": {
      "patch": {
        "summary": "Rename File",
        "description": "Renames a file.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File renamed successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "summary": "Delete File",
        "description": "Deletes a file.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/files/preview": {
      "get": {
        "summary": "Preview File",
        "description": "Returns a preview of a file using a valid token.",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File preview returned successfully."
          },
          "400": {
            "description": "Missing or invalid token."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  "/v1/files/generate-token": {
    "post": {
      "summary": "Generate File Access Token",
      "description": "Generates a secure token for file preview/download.",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "fileId": {
                  "type": "string"
                },
                "expiresIn": {
                  "type": "string"
                }
              },
              "required": [
                "fileId"
              ]
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Token generated successfully."
        },
        "400": {
          "description": "Validation error."
        },
        "403": {
          "description": "Forbidden."
        },
        "404": {
          "description": "File not found."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/files/download": {
    "get": {
      "summary": "Download File",
      "description": "Downloads a file using a valid token.",
      "parameters": [
        {
          "name": "token",
          "in": "query",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "File downloaded successfully."
        },
        "400": {
          "description": "Missing or invalid token."
        },
        "404": {
          "description": "File not found."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
...
}
}